% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grreat.R
\name{grreat}
\alias{grreat}
\title{Praise food}
\usage{
grreat(food)
}
\arguments{
\item{food}{String naming a food}
}
\value{

}
\description{
Print a comment to the console that praises the food you specify like Tony the Tiger.
}
\examples{
grreat(food = "Okra")
}
